refactor:#84-DeatilCard컴포넌트화
